+++ libraries/session/src/main/java/androidx/media3/session/MediaController.java
@@ -201,6 +201,8 @@ public class MediaController implements Player {
       "MediaController method is called from a wrong thread."
           + " See javadoc of MediaController for details.";
 
+  @UnstableApi protected static final long DEFAULT_PLATFORM_CALLBACK_AGGREGATION_TIMEOUT_MS = 100L;
+
   /** A builder for {@link MediaController}. */
   public static final class Builder {
 
@@ -211,6 +213,7 @@ public class MediaController implements Player {
     private Looper applicationLooper;
     private @MonotonicNonNull BitmapLoader bitmapLoader;
     private int maxCommandsForMediaItems;
+    private long platformSessionCallbackAggregationTimeoutMs;
 
     /**
      * Creates a builder for {@link MediaController}.
@@ -242,6 +245,8 @@ public class MediaController implements Player {
       connectionHints = Bundle.EMPTY;
       listener = new Listener() {};
       applicationLooper = Util.getCurrentOrMainLooper();
+      platformSessionCallbackAggregationTimeoutMs =
+          DEFAULT_PLATFORM_CALLBACK_AGGREGATION_TIMEOUT_MS;
     }
 
     /**
@@ -321,6 +326,24 @@ public class MediaController implements Player {
     }
 
     /**
+     * Sets the timeout after which updates from the platform session callbacks are applied to the
+     * browser, in milliseconds.
+     *
+     * <p>The default is 100ms.
+     *
+     * @param platformSessionCallbackAggregationTimeoutMs The timeout, in milliseconds.
+     * @return tThe builder to allow chaining.
+     */
+    @UnstableApi
+    @CanIgnoreReturnValue
+    public Builder experimentalSetPlatformSessionCallbackAggregationTimeoutMs(
+        long platformSessionCallbackAggregationTimeoutMs) {
+      this.platformSessionCallbackAggregationTimeoutMs =
+          platformSessionCallbackAggregationTimeoutMs;
+      return this;
+    }
+
+    /**
      * Builds a {@link MediaController} asynchronously.
      *
      * <p>The controller instance can be obtained like the following example:
@@ -361,7 +384,8 @@ public class MediaController implements Player {
               applicationLooper,
               holder,
               bitmapLoader,
-              maxCommandsForMediaItems);
+              maxCommandsForMediaItems,
+              platformSessionCallbackAggregationTimeoutMs);
       postOrRun(new Handler(applicationLooper), () -> holder.setController(controller));
       return holder;
     }
@@ -553,7 +577,8 @@ public class MediaController implements Player {
       Looper applicationLooper,
       ConnectionCallback connectionCallback,
       @Nullable BitmapLoader bitmapLoader,
-      int maxCommandsForMediaItems) {
+      int maxCommandsForMediaItems,
+      long platformSessionCallbackAggregationTimeoutMs) {
     checkNotNull(context, "context must not be null");
     checkNotNull(token, "token must not be null");
     Log.i(
@@ -576,7 +601,14 @@ public class MediaController implements Player {
     this.connectionCallback = connectionCallback;
     this.maxCommandsForMediaItems = maxCommandsForMediaItems;
 
-    impl = createImpl(context, token, connectionHints, applicationLooper, bitmapLoader);
+    impl =
+        createImpl(
+            context,
+            token,
+            connectionHints,
+            applicationLooper,
+            bitmapLoader,
+            platformSessionCallbackAggregationTimeoutMs);
     impl.connect();
   }
 
@@ -587,10 +619,17 @@ public class MediaController implements Player {
       SessionToken token,
       Bundle connectionHints,
       Looper applicationLooper,
-      @Nullable BitmapLoader bitmapLoader) {
+      @Nullable BitmapLoader bitmapLoader,
+      long platformSessionCallbackAggregationTimeoutMs) {
     if (token.isLegacySession()) {
       return new MediaControllerImplLegacy(
-          context, this, token, connectionHints, applicationLooper, checkNotNull(bitmapLoader));
+          context,
+          this,
+          token,
+          connectionHints,
+          applicationLooper,
+          checkNotNull(bitmapLoader),
+          platformSessionCallbackAggregationTimeoutMs);
     } else {
       return new MediaControllerImplBase(context, this, token, connectionHints, applicationLooper);
     }